+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/ForwardingAudioSink.java
@@ -173,8 +173,15 @@ public class ForwardingAudioSink implements AudioSink {
   }
 
   @Override
+  @RequiresApi(29)
   public void setOffloadMode(@OffloadMode int offloadMode) {
     sink.setOffloadMode(offloadMode);
+  }
+
+  @Override
+  @RequiresApi(29)
+  public void setOffloadDelayPadding(int delayInFrames, int paddingInFrames) {
+    sink.setOffloadDelayPadding(delayInFrames, paddingInFrames);
   }
 
   @Override